Update security, business logic, performance and add SQL schema changes - 27/04/2026

+++ src/MotoShop/Controllers/PaymentController.cs
@@ -46,7 +46,7 @@ namespace MotoShop.Controllers
             vnpay.AddRequestData("vnp_Version", "2.1.0");
             vnpay.AddRequestData("vnp_Command", "pay");
             vnpay.AddRequestData("vnp_TmnCode", vnp_TmnCode);
-            vnpay.AddRequestData("vnp_Amount", ((int)order.TotalAmount * 100).ToString()); 
+            vnpay.AddRequestData("vnp_Amount", ((long)(order.TotalAmount * 100)).ToString()); 
             vnpay.AddRequestData("vnp_CreateDate", DateTime.Now.ToString("yyyyMMddHHmmss"));
             vnpay.AddRequestData("vnp_CurrCode", "VND");
             vnpay.AddRequestData("vnp_IpAddr", HttpContext.Connection.RemoteIpAddress?.ToString() ?? "127.0.0.1");
